Implemented block-threaded rendering (one thread per image block) and block update flashes for eye candy.
--HG--
extra : convert_revision : svn%3Ac32f086a-3a94-11de-9146-fdf1833cc731/trunk%40395

+++ src/tracer/path.cpp
@@ -25,7 +25,7 @@ const rgbColor pathTracer::_L(ray& r, const int depth) const {
 
         if(!isect.hit){
             if(pathLength == 0 || lastBounceWasSpecular){
-                for(int i=0; i<parent.numLights(); ++i){
+                for(size_t i=0; i<parent.numLights(); ++i){
                     const float lightDist = (parent.getLight(i)->getPosition() - rOrig.origin).length();
                     ray lightRay(rOrig, EPSILON, lightDist);
